requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
dj-database-url==1.2.0
Django==4.1.7
django-filter==22.1
django-multiselectfield==0.1.12
django-redis==5.2.0
djangorestframework==3.14.0    
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.2.2
mapscript==8.0.0
Markdown==3.4.1
Pillow==9.4.0
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.5.4
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.4.0